requirements.txt
-i https://pypi.org/simple
altair==5.4.1; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flatbuffers==24.3.25
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.8; python_version >= '3.6'
imageio==2.35.1; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
lazy-loader==0.4; python_version >= '3.7'
llvmlite==0.43.0; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
mpmath==1.3.0
narwhals==1.6.0; python_version >= '3.8'
networkx==3.3; python_version >= '3.10'
numba==0.60.0; python_version >= '3.9'
numpy==2.0.2; python_version >= '3.9'
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pillow==10.4.0; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
pooch==1.8.2; python_version >= '3.7'
protobuf==5.28.0; python_version >= '3.8'
pyarrow==17.0.0; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pymatting==1.1.12; python_version >= '3'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
referencing==0.35.1; python_version >= '3.8'
rembg==2.0.59; python_version < '3.13' and python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==13.8.0; python_full_version >= '3.7.0'
rpds-py==0.20.0; python_version >= '3.8'
scikit-image==0.24.0; python_version >= '3.9'
scipy==1.14.1; python_version >= '3.10'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
streamlit==1.38.0; python_version >= '3.8' and python_full_version != '3.9.7'
sympy==1.13.2; python_version >= '3.8'
tenacity==8.5.0; python_version >= '3.8'
tifffile==2024.8.30; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.4.1; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
